refactor: share the upload card and staged file list across import flows
