+++ frontend/src/pages/imports/components/index.ts
@@ -1,3 +1,4 @@
+export * from './FileUpload'
 export * from './tables/AccountMappingTable'
 export * from './tables/HeaderMappingTable'
 export * from './tables/CategoryTypeToggle'
